added back data into json file, added pages dir and put home page into its own page
