+++ components/Table.vue
@@ -27,28 +27,53 @@ export default {
       data: this.tableData, //link data to table
       reactiveData: true, //enable data reactivity
       columns: [
-      { title: "companyUrl", field: "companyUrl", sorter: "string", width: 200 },
-      { title: "companyName", field: "companyName", sorter: "string" },
-      { title: "jobTitle", field: "jobTitle", sorter: "string", width: 200 },
-      // { title: "jobLocation", field: "jobLocation", sorter: "string", width: 200 },
-      // { title: "postedAt", field: "postedAt", sorter: "date", width: 200 },
-      // { title: "appliesClosedAt", field: "appliesClosedAt", sorter: "date", width: 200 },
-      // { title: "jobDescription", field: "jobDescription", sorter: "string", width: 200 },
-      // { title: "jobId", field: "jobId", sorter: "number", width: 200 },
-      // { title: "workplaceType", field: "workplaceType", sorter: "string", width: 200 },
-      // { title: "jobPosterProfileUrl", field: "jobPosterProfileUrl", sorter: "string", width: 200 },
-      // { title: "jobPosterName", field: "jobPosterName", sorter: "string", width: 200 },
-      // { title: "companyLogoUrl", field: "companyLogoUrl", sorter: "string", width: 200 },
-      // { title: "applyUrl", field: "applyUrl", sorter: "string", width: 200 },
-      // { title: "viewsCount", field: "viewsCount", sorter: "number", width: 200 },
-      // { title: "companyStaffCount", field: "companyStaffCount", sorter: "number", width: 200 },
-      // { title: "companyDescription", field: "companyDescription", sorter: "string", width: 200 },
-      // { title: "jobIndustries", field: "jobIndustries", sorter: "string", width: 200 },
-      // { title: "jobFunctions", field: "jobFunctions", sorter: "string", width: 200 },
-      // { title: "remoteAllowed", field: "remoteAllowed", sorter: "boolean", width: 200 },
-      // { title: "jobType", field: "jobType", sorter: "string", width: 200 },
-      // { title: "applicantsCount", field: "applicantsCount", sorter: "number", width: 200 },
-      // { title: "experienceLevel", field: "experienceLevel", sorter: "string", width: 200 },
+      { title: "Job ID", field: "jobId", sorter: "number", width: 100 },
+      { title: "Job Title", field: "jobTitle", sorter: "string", width: 200},
+      { title: "Company Name", field: "companyName", sorter: "string", width: 100 },
+      { title: "Company URL", field: "companyUrl", sorter: "string", width: 200, visible: false },
+      { title: "Job Location", field: "jobLocation", sorter: "string", width: 120 },
+      { title: "Posted At", field: "postedAt", sorter: "date", width: 200, visible: false },
+      { title: "Applies Closed At", field: "appliesClosedAt", sorter: "date", width: 200, visible: false },
+      { title: "Job Description", field: "jobDescription", sorter: "string", width: 200, visible: false },
+      { title: "Workplace Type", field: "workplaceType", sorter: "string", width: 100, visible: false  },
+      { title: "Job Poster Profile URL", field: "jobPosterProfileUrl", sorter: "string", width: 200, visible: false },
+      { title: "Job Poster Name", field: "jobPosterName", sorter: "string", width: 100 },
+      { title: "Company Logo URL", field: "companyLogoUrl", sorter: "string", width: 200, visible: false },
+      { title: "Apply URL", field: "applyUrl", sorter: "string", width: 200, visible: false },
+      { title: "Views Count", field: "viewsCount", sorter: "number", width: 80, visible: false },
+      { title: "Company Staff Count", field: "companyStaffCount", sorter: "number", width: 200, visible: false },
+      { title: "Company Description", field: "companyDescription", sorter: "string", width: 200, visible: false },
+      { title: "Job Industries", field: "jobIndustries", sorter: "string", width: 200, visible: false  },
+      { title: "Job Functions", field: "jobFunctions", sorter: "string", width: 200, visible: false },
+      { title: "Remote Allowed", field: "remoteAllowed", sorter: "boolean", width: 80 },
+      { title: "Job Type", field: "jobType", sorter: "string", width: 80 },
+      { title: "Applicants Count", field: "applicantsCount", sorter: "number", width: 80, visible: false },
+      { title: "Experience Level", field: "experienceLevel", sorter: "string", width: 100 },
+       {
+          title: "Actions",
+          field: "actions",
+          formatter: function (cell, formatterParams, onRendered) {
+            // Create and style button
+            let button = document.createElement("button");
+            button.style.padding = "10px";
+            button.style.backgroundColor = "lightblue";
+            button.style.border = "none";
+            button.style.borderRadius = "5px";
+            button.style.cursor = "pointer";
+            // button.innerHTML = cell.getValue();
+            button.innerHTML = 'Send Email';
+
+            // Add event listener
+            button.addEventListener("click", (e) => {
+              // Emit an event with the row data
+              vueInstance.$emit('send-email', cell.getRow().getData());
+              console.log("Button clicked:", cell.getRow().getData());
+              console.log("getData", cell.getRow());
+            });
+
+            return button;
+          }
+        },
       ],
     });
     this.tabulator.on("rowDblClick", function (e, row) {
