+++ pages/job.vue
@@ -1,0 +1,3 @@
+<template>
+  <div>Job page</div>
+</template>
