+++ server/api/hello.ts
@@ -1,0 +1,6 @@
+import data from './jobs.json';
+export default defineEventHandler((event) => {
+  return {
+    data,
+  }
+});
